Split map page into files
