+++ lib/src/models/rover/rover.dart
@@ -7,7 +7,7 @@ import "package:rover_dashboard/models.dart";
 import "settings.dart";
 
 /// The model to control the entire rover.
-/// 
+///
 /// Find more specific functionality in this class's fields.
 class Rover extends Model {
 	/// Monitors metrics coming from the rover.
@@ -41,13 +41,13 @@ class Rover extends Model {
   Iterable<Controller> get controllers => [controller1, controller2, controller3];
 
 	/// Whether the rover is connected.
-	bool get isConnected => models.sockets.data.isConnected;
+	bool get isConnected => models.sockets.sockets.any((socket) => socket.isConnected);
 
 	/// The current status of the rover.
 	ValueNotifier<RoverStatus> status = ValueNotifier(RoverStatus.DISCONNECTED);
 
 	@override
-	Future<void> init() async { 
+	Future<void> init() async {
     setDefaultControls();
     await metrics.init();
 		await controller1.init();
